implement FST + FSA intersection that leverages fast addressing of arc/transitions
FST nodes have differetn variant. For non-variable length encoded node we can more efficiently
lookup for a target label.

Similarly, for FSAs the TransitionAccessor allows access to a list of [min, max] ranges in order, on which
we can perform binary-search to advance to applicable transitions for a given target